+++ user_guide_src/source/general/context/011.php
@@ -1,5 +1,3 @@
 <?php
 
-if ($context->missing('user_id')) {
-    // user_id hasn't been set yet
-}
+$context->remove('user_id');
