+++ user_guide_src/source/general/context/013.php
@@ -1,3 +1,3 @@
 <?php
 
-$context->remove(['user_id', 'username', 'request_id']);
+$context->clear();
