+++ user_guide_src/source/general/context/014.php
@@ -1,3 +1,3 @@
 <?php
 
-$context->clear();
+$context->setHidden('api_key', 'sk_live_abc123xyz789');
