+++ user_guide_src/source/general/context/015.php
@@ -1,3 +1,7 @@
 <?php
 
-$context->setHidden('api_key', 'sk_live_abc123xyz789');
+$context->setHidden([
+    'api_key'     => 'sk_live_abc123xyz789',
+    'api_secret'  => 'secret_key_here',
+    'db_password' => 'database_password',
+]);
