+++ user_guide_src/source/general/context/016.php
@@ -1,7 +1,4 @@
 <?php
 
-$context->setHidden([
-    'api_key'     => 'sk_live_abc123xyz789',
-    'api_secret'  => 'secret_key_here',
-    'db_password' => 'database_password',
-]);
+$apiKey = $context->getHidden('api_key');
+// $apiKey = 'sk_live_abc123xyz789'
