+++ user_guide_src/source/general/context/017.php
@@ -1,4 +1,13 @@
 <?php
 
-$apiKey = $context->getHidden('api_key');
-// $apiKey = 'sk_live_abc123xyz789'
+// Get with default value
+$apiKey = $context->getHidden('api_key', 'default_key');
+
+// Get only specific hidden keys
+$credentials = $context->getOnlyHidden(['api_key', 'api_secret']);
+
+// Get all hidden data except specific keys
+$data = $context->getExceptHidden(['db_password']);
+
+// Get all hidden data
+$allHidden = $context->getAllHidden();
