+++ user_guide_src/source/general/context/020.php
@@ -1,7 +1,3 @@
 <?php
 
-// Remove a single hidden value
-$context->removeHidden('api_key');
-
-// Remove multiple hidden values
-$context->removeHidden(['api_key', 'api_secret']);
+$context->clearHidden();
