+++ user_guide_src/source/general/context/021.php
@@ -1,3 +1,3 @@
 <?php
 
-$context->clearHidden();
+$context->clearAll();
