+++ user_guide_src/source/general/context/022.php
@@ -1,3 +1,14 @@
 <?php
 
-$context->clearAll();
+namespace Config;
+
+use CodeIgniter\Config\BaseConfig;
+
+class Logger extends BaseConfig
+{
+    // ...
+
+    public bool $logGlobalContext = true;
+
+    // ...
+}
