--- user_guide_src/source/general/context/024.php
@@ -1,7 +1,0 @@
-<?php
-
-$context = service('context');
-$context->set('user_id', 123);
-$context->set('transaction_id', 'txn_12345');
-
-log_message('error', 'Payment processing failed');
